Add new 'Runtime' column for jobs
This is calculated by Pulpito as the difference between the posted and
updated times of a job. This will often differ wildly from the
'Duration' column, as that is reported by teuthology and is determined
by the start and end times of the timer task. In short, 'runtime' is the
amount of time a job held up a worker process and 'duration' is the
amount of time a job was actually executing a test.