Add Dockerfile to build pulpito inside a container
Add a Dockerfile to build pulpito inside a container

This is useful to be able to run pulpito with docker or podman.
The container can be build with:

podman build -f Dockerfile

and then started with:

podman run -p 80:8080 -e PULPITO_PADDLES_ADDRESS=http://paddles:1234 \
    $IMAGE_ID

Where $IMAGE_ID is the podman image id from the previous image build
run. PULPITO_PADDLES_ADDRESS is the address of the paddles server.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>